External database
I have modified the app to work on an external database. I set up an
RDS instance on amazon, and am directing traffic there. (Success) App
still routes traffic back to host machine, aka laptop.

Next step is to upload the app and make some changes so that it can
actually run from a website, to mimic usability in the real world.